20251109_EdgeTAM/pyproject.toml
[project]
name = "edgetam-project"
version = "0.1.0"
description = "EdgeTAM - On-Device Track Anything Model"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.40.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "timm>=0.9.0",
]
